[MOD] cambios ultima hora

+++ ProyectoFinalAngular/src/app/inyectador/modificador/modificador.component.ts
@@ -1,8 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
+import { DialogoConfirmacionComponent } from 'src/app/modal/dialogo-confirmacion/dialogo-confirmacion.component';
 import { PeticionesService } from 'src/app/servicios/peticiones.service';
 
 @Component({
@@ -15,7 +17,7 @@ export class ModificadorComponent implements OnInit {
   formulario: FormGroup;
   productos: any = [];
   page=1;
-  constructor(private router: Router, private peticiones: PeticionesService, private http: HttpClient,private toastr: ToastrService) { }
+  constructor(private router: Router, private peticiones: PeticionesService, private http: HttpClient,private toastr: ToastrService,public matDialog: MatDialog) { }
   
   
 
@@ -45,11 +47,22 @@ export class ModificadorComponent implements OnInit {
     this.router.navigate(["anadir-componente"])
   }
  borrarProducto(id){
-  this.peticiones.deleteProducto(id).subscribe(data=>{
-      this.getProductos();
-  },error=>{
-    this.toastr.error("No se ha podido eliminar el producto");
+
+  this.matDialog
+  .open(DialogoConfirmacionComponent, {
+    data: `¿Esta seguro de borrar el producto? ` 
   })
-  this.toastr.info("Se ha eliminado el producto con exito");
- }
+  .afterClosed()
+  .subscribe((confirmado: Boolean) => {
+    if (confirmado) {
+      this.toastr.info("Se ha eliminado el producto con exito");
+      this.peticiones.deleteProducto(id).subscribe(data=>{
+      },error=>{
+        this.toastr.error("No se ha podido eliminar el producto");
+      })
+    } 
+  });
+ 
+}
+ 
 }
